+++ R/get_index.R
@@ -10,19 +10,12 @@
 get_index <- function(data, filter = NULL, scale = NULL, subscale = NULL,
                       subscale_2 = NULL, names = TRUE, class = NULL) {
 
+  warning("get_index() is deprecated. Please use function select_items instead.")
   filter <- deparse(substitute(filter))
   if (!is.null(scale) || !is.null(subscale) || !is.null(subscale_2)) {
     filter <- .to_filter(scale = scale, subscale = subscale, subscale_2 = subscale_2)
   }
 
-    .get_index(data = data, filter = filter, names = names, class = NULL)
+    .get_index(data = data, filter = filter, names = names, class = class)
 }
 
-.get_index <- function(data, filter, names = TRUE, class = NULL) {
-  if (!is.null(class)) filter <- paste0(filter, " & class == '", class, "'")
-  dic <- extract_dic(data)
-  id <- with(dic, eval(str2lang(filter)))
-  out <- which(names(data) %in% dic[[.opt$item_name]][id])
-  if (names) out <- names(data)[out]
-  out
-}