+++ R/rename_by_list.R
@@ -45,6 +45,12 @@ rename_by_list <- function(data,
   rn <- setNames(from, to)
   rn <- rn[rn %in% names(data)]
   rn <- rn[!is.na(names(rn))]
+
+  if (length(rn) == 0) stop("No variables renamed")
+
   out <- rename(data, !!rn)
+
+  cat(paste0(length(rn), " variables renamed\n"))
+
   out
 }