+++ scaledic_data.RData.R
@@ -1,7 +1,6 @@
-library(openxlsx)
 
-dic_ITRF <- read.xlsx("dic_itrf.xlsx")
-ITRF <- read.xlsx("itrf.xlsx")
+dic_ITRF <- openxlsx::read.xlsx("dic_itrf.xlsx")
+ITRF <- openxlsx::read.xlsx("itrf.xlsx")
 
 save(ITRF, dic_ITRF, file = "data/scaledic.RData")